+++ packages/gatsby/src/services/post-bootstrap.ts
@@ -1,8 +1,8 @@
 import reporter from "gatsby-cli/lib/reporter"
-import { emitter } from "../redux"
+import { emitter, store } from "../redux"
 import apiRunnerNode from "../utils/api-runner-node"
 import { IDataLayerContext } from "../state-machines/data-layer/types"
-import { boundActionCreators } from "../redux/actions"
+import { actions } from "../redux/actions"
 
 export async function postBootstrap({
   parentSpan,
@@ -22,5 +22,5 @@ export async function postBootstrap({
     bootstrap finished - ${process.uptime().toFixed(3)}s
   `)
   emitter.emit(`BOOTSTRAP_FINISHED`, {})
-  boundActionCreators.setProgramStatus(`BOOTSTRAP_FINISHED`)
+  store.dispatch(actions.setProgramStatus(`BOOTSTRAP_FINISHED`))
 }