+++ packages/gatsby/src/utils/worker/child.ts
@@ -1,2 +1,2 @@
 // Note: this doesn't check for conflicts between module exports
-export { renderHTML } from "./render-html"
+export { renderHTMLProd, renderHTMLDev } from "./render-html"